Handle user creation with and without limit errors in User Management page alerts
