+++ packages/components/src/internal/components/security/APIWrapper.ts
@@ -1,48 +1,23 @@
-import {ActionURL, Ajax, Security, Utils} from '@labkey/api';
+import { Security } from '@labkey/api';
 import { Map } from 'immutable';
 
 import { Container } from '../base/models/Container';
-import { fetchContainerSecurityPolicy } from '../permissions/actions';
+import { fetchContainerSecurityPolicy, UserLimitSettings, getUserLimitSettings } from '../permissions/actions';
 import { Principal, SecurityPolicy } from '../permissions/models';
 
 export type FetchContainerOptions = Omit<Security.GetContainersOptions, 'success' | 'failure' | 'scope'>;
 
-export type UserLimitSettings = {
-    activeUsers: number;
-    messageHtml: string;
-    remainingUsers: number;
-    success: boolean;
-    userLimitLevel: number;
-    userLimit: boolean;
-};
-
 export interface SecurityAPIWrapper {
-    getUserLimitSettings: () => Promise<UserLimitSettings>;
     fetchContainers: (options: FetchContainerOptions) => Promise<Container[]>;
     fetchPolicy: (
         containerId: string,
         principalsById: Map<number, Principal>,
         inactiveUsersById?: Map<number, Principal>
     ) => Promise<SecurityPolicy>;
+    getUserLimitSettings: () => Promise<UserLimitSettings>;
 }
 
 export class ServerSecurityAPIWrapper implements SecurityAPIWrapper {
-    getUserLimitSettings = (): Promise<UserLimitSettings> => {
-        return new Promise((resolve, reject) => {
-            Ajax.request({
-                url: ActionURL.buildURL('user', 'getuserLimitSettings.api'),
-                method: 'GET',
-                scope: this,
-                success: Utils.getCallbackWrapper(settings => {
-                    resolve(settings as UserLimitSettings);
-                }),
-                failure: Utils.getCallbackWrapper(error => {
-                    console.error(error);
-                    reject(error);
-                }),
-            });
-        });
-    };
     fetchContainers = (options: FetchContainerOptions): Promise<Container[]> => {
         return new Promise((resolve, reject) => {
             Security.getContainers({
@@ -58,6 +33,7 @@ export class ServerSecurityAPIWrapper implements SecurityAPIWrapper {
         });
     };
     fetchPolicy = fetchContainerSecurityPolicy;
+    getUserLimitSettings = getUserLimitSettings;
 }
 
 function recurseContainerHierarchy(data: Security.ContainerHierarchy, container: Container): Container[] {
@@ -75,9 +51,9 @@ export function getSecurityTestAPIWrapper(
     overrides: Partial<SecurityAPIWrapper> = {}
 ): SecurityAPIWrapper {
     return {
-        getUserLimitSettings: mockFn(),
         fetchContainers: mockFn(),
         fetchPolicy: mockFn(),
+        getUserLimitSettings: mockFn(),
         ...overrides,
     };
 }
